+++ src/main/java/org/realityforge/gwt/webpoller/server/AbstractJaxrsHttpRequestFactory.java
@@ -115,7 +115,7 @@ public abstract class AbstractJaxrsHttpRequestFactory
     }
 
     @Override
-    public void cancel()
+    public synchronized void cancel()
     {
       if ( null != _future )
       {