+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Move to Java 8 and GWT 2.8.
 * Add a `ReentrantReadWriteLock` around access to timer in `server.TimerBasedWebPoller`.
+* Synchronize access to Request.cancel() method in server mode to avoid doubly canceling futures.
 
 ## 0.9.2: