+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.9.3 (Pending):
+## 0.9.3:
 
 * Move to Java 8 and GWT 2.8.
 * Add a `ReentrantReadWriteLock` around access to critical resources in `server.TimerBasedWebPoller`.